-some thread safety -reduce allocations -additional 64bit counters to avoid overflows
git-svn-id: http://svn.code.sf.net/p/azsmrc/code/mldht/trunk@2052 58cabe8c-710f-0410-8395-bf4839b10e17

+++ lbms/plugins/mldht/kad/DHT.java
@@ -866,7 +866,7 @@ public class DHT implements DHTBase {
 	private void onStatsUpdate () {
 		stats.setNumTasks(tman.getNumTasks() + tman.getNumQueuedTasks());
 		stats.setNumPeers(node.getNumEntriesInRoutingTable());
-		int numSent = 0;int numReceived = 0;int activeCalls = 0;
+		long numSent = 0;long numReceived = 0;int activeCalls = 0;
 		for(RPCServer s : serverManager.getAllServers())
 		{
 			numSent += s.getNumSent();
